Move custom fields to workspace settings and use case breadcrumbs
