+++ frontend/src/app/workspaces/[workspaceId]/cases/layout.tsx
@@ -1,9 +1,10 @@
 "use client"
 
+import { format } from "date-fns"
+import { CirclePlusIcon } from "lucide-react"
 import Link from "next/link"
 import { usePathname, useSearchParams } from "next/navigation"
 import { Suspense } from "react"
-import { CaseTableInsertButton } from "@/components/cases/case-table-manage-button"
 import { CenteredSpinner } from "@/components/loading/spinner"
 import {
   Breadcrumb,
@@ -12,6 +13,8 @@ import {
   BreadcrumbList,
   BreadcrumbSeparator,
 } from "@/components/ui/breadcrumb"
+import { Button } from "@/components/ui/button"
+import { useCreateCase } from "@/lib/hooks"
 import { cn } from "@/lib/utils"
 import { useWorkspace } from "@/providers/workspace"
 
@@ -20,7 +23,7 @@ function BreadcrumbNavigation() {
   const searchParams = useSearchParams()
   const pathname = usePathname()
   const queryCategory = searchParams.get("category")
-  const isFieldsPage = pathname.endsWith("/cases/fields")
+  const { createCase, createCaseIsPending } = useCreateCase(workspaceId)
 
   // Check if we're on a case detail page
   const pathSegments = pathname.split("/")
@@ -28,9 +31,18 @@ function BreadcrumbNavigation() {
   const isCaseDetailPage =
     caseIndex !== -1 &&
     pathSegments[caseIndex + 1] &&
-    pathSegments[caseIndex + 1] !== "fields" &&
     pathSegments[caseIndex + 1].match(/^[a-zA-Z0-9-]+$/)
   const caseId = isCaseDetailPage ? pathSegments[caseIndex + 1] : null
+
+  const handleCreateCase = () => {
+    createCase({
+      summary: `New case - ${format(new Date(), "PPpp")}`,
+      description: "",
+      status: "unknown",
+      priority: "unknown",
+      severity: "unknown",
+    })
+  }
 
   return (
     <div className="flex w-full items-center justify-between">
@@ -42,12 +54,10 @@ function BreadcrumbNavigation() {
                 asChild
                 className={cn(
                   "flex items-center",
-                  queryCategory || isFieldsPage || isCaseDetailPage
+                  queryCategory || isCaseDetailPage
                     ? "text-muted-foreground"
                     : "text-foreground",
-                  !isFieldsPage &&
-                    !isCaseDetailPage &&
-                    "cursor-default text-foreground"
+                  !isCaseDetailPage && "cursor-default text-foreground"
                 )}
               >
                 <Link href={`/workspaces/${workspaceId}/cases`}>
@@ -57,18 +67,6 @@ function BreadcrumbNavigation() {
                 </Link>
               </BreadcrumbLink>
             </BreadcrumbItem>
-            {isFieldsPage && (
-              <>
-                <BreadcrumbSeparator>{"/"}</BreadcrumbSeparator>
-                <BreadcrumbItem>
-                  <BreadcrumbLink>
-                    <span className="text-2xl font-semibold tracking-tight text-foreground/80">
-                      Custom fields
-                    </span>
-                  </BreadcrumbLink>
-                </BreadcrumbItem>
-              </>
-            )}
             {isCaseDetailPage && caseId && (
               <>
                 <BreadcrumbSeparator>{"/"}</BreadcrumbSeparator>
@@ -85,7 +83,16 @@ function BreadcrumbNavigation() {
         </Breadcrumb>
       </div>
       <div className="ml-auto flex items-center space-x-2">
-        {!isCaseDetailPage && <CaseTableInsertButton />}
+        {!isCaseDetailPage && (
+          <Button
+            onClick={handleCreateCase}
+            disabled={createCaseIsPending}
+            className="h-7 items-center space-x-1 bg-emerald-500/80 px-3 py-1 text-xs text-white shadow-sm hover:border-emerald-500 hover:bg-emerald-400/80"
+          >
+            <CirclePlusIcon className="size-3" />
+            <span>Open case</span>
+          </Button>
+        )}
       </div>
     </div>
   )
