+++ frontend/src/app/workspaces/[workspaceId]/settings/custom-fields/layout.tsx
@@ -1,0 +1,13 @@
+import type { Metadata } from "next"
+
+export const metadata: Metadata = {
+  title: "Custom fields | Workspace",
+}
+
+export default function CustomFieldsLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return children
+}
